multi_doc_chat/model/models.py
from pydantic import BaseModel,Field
from typing import Annotated
from enum import Enum


class ChatAnswer(BaseModel):
    """Validate chat answer type and length"""
    answer: Annotated[str,Field(min_length=1,max_length=2048)]

class PromptType(str,Enum):
    CONTEXTUALIZE_QUESTION = "contextualize_question"
    CONTEXT_QA = "context_qa"

class UploadResume(BaseModel):
    session_id: str
    indexed:bool
    message: str | None=None

class ChatRequest(BaseModel):
    session_id:str
    message:str

class ChatResponse:
    answer:str