pyproject.toml
[project]
name = "fully-deploable-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docx2txt>=0.9",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "ipykernel>=7.2.0",
    "jinja2>=3.1.6",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.17",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-nebius>=0.1.3",
    "langchain-openai>=1.1.10",
    "langchain-text-splitters>=1.1.1",
    "langsmith>=0.7.14",
    "openai>=2.26.0",
    "pandas>=3.0.1",
    "pypdf>=6.7.5",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "structlog>=25.5.0",
    "uvicorn>=0.41.0",
]
